поменял кучу всего не удалось вызвать доставание бина с параметрами

+++ Calc/src/intf/Viewer.java
@@ -2,6 +2,6 @@ package intf;
 
 public interface Viewer {
 
-	public  void onView(Integer result);
+	public  void onView(Double result);
 	
 }
