+++ Calc/src/viewers/SuperConsoleView.java
@@ -5,7 +5,7 @@ import intf.Viewer;
 public class SuperConsoleView implements Viewer {
 
 	@Override
-	public void onView(Integer result) {
+	public void onView(Double result) {
 		System.out.println("the super puper sping framework, \n and result is :" + result);
 
 	}
